Update tick.h: ticks are 32 bits

+++ src/network_messages/tick.h
@@ -140,6 +140,6 @@ struct CurrentTickInfo
     unsigned int tick;
     unsigned short numberOfAlignedVotes;
     unsigned short numberOfMisalignedVotes;
-    unsigned short initialTick;
+    unsigned int initialTick;
 };
